SwapAny: updated structs and hashing; initial unit test
