+++ tools/constants/index.ts
@@ -1,5 +1,7 @@
 export const DOMAIN_NAME_SWAP = 'SWAP'
 export const DOMAIN_VERSION_SWAP = '3'
+export const DOMAIN_NAME_SWAP_ANY = 'SWAP_ANY'
+export const DOMAIN_VERSION_SWAP_ANY = '1'
 export const DOMAIN_NAME_POOL = 'POOL'
 export const DOMAIN_VERSION_POOL = '1'
 export const ADDRESS_ZERO = '0x0000000000000000000000000000000000000000'
